#200 - 金額の単位を設定可能にする
- 幾つかの金額に、デフォルトに設定したアセットの単位(前置/後置)を表示するように変更
- グラフの縦軸の表記にデフォルトに設定したアセットの名称を表示するように変更

+++ HouseholdAccountBook/Infrastructure/CSV/ActionCsvDto.cs
@@ -14,9 +14,9 @@ namespace HouseholdAccountBook.Infrastructure.CSV
         public DateTime Date { get; init; }
 
         /// <summary>
-        /// 値
+        /// 値(主単位)
         /// </summary>
-        public int Value { get; init; }
+        public decimal Value { get; init; }
 
         /// <summary>
         /// 名称
